requirements.txt
a2a-sdk==0.3.0
dotenv==0.9.9
google-adk==1.9.0
google-api-python-client==2.178.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.107.0
google-cloud-discoveryengine==0.13.11
google-cloud-logging==3.12.1
google-cloud-secret-manager==2.24.0
google-generativeai==0.8.5
Pillow==11.3.0
stackapi==0.3.1
uvicorn==0.35.0
